test : LottoWinner rank
